requirements.txt
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
numpy==2.1.3
opencv-python==4.10.0.84
pandas==2.2.3
pillow==11.0.0
python-json-logger==2.0.7
torch==2.5.1+cu118
torchvision==0.20.1+cu118
tzdata==2024.2
ultralytics==8.3.38
ultralytics-thop==2.0.12